refactor: split UI variants and utilities
